--- yudao-framework/yudao-spring-boot-starter-biz-tenant/src/main/java/cn/iocoder/yudao/framework/tenant/core/job/TenantJobHandlerInvoker.java
@@ -1,14 +1,0 @@
-package cn.iocoder.yudao.framework.tenant.core.job;
-
-import cn.iocoder.yudao.framework.quartz.core.handler.JobHandlerInvoker;
-
-/**
- * 多租户 JobHandlerInvoker 拓展实现类
- *
- * @author 芋道源码
- */
-public class TenantJobHandlerInvoker extends JobHandlerInvoker {
-
-
-
-}
