+++ yudao-module-infra/yudao-module-infra-impl/src/main/java/cn/iocoder/yudao/module/infra/dal/mysql/file/FileMapper.java
@@ -5,7 +5,6 @@ import cn.iocoder.yudao.framework.mybatis.core.mapper.BaseMapperX;
 import cn.iocoder.yudao.framework.mybatis.core.query.QueryWrapperX;
 import cn.iocoder.yudao.module.infra.controller.admin.file.vo.FilePageReqVO;
 import cn.iocoder.yudao.module.infra.dal.dataobject.file.FileDO;
-import com.baomidou.mybatisplus.annotation.InterceptorIgnore;
 import org.apache.ibatis.annotations.Mapper;
 
 /**
@@ -31,12 +30,10 @@ public interface FileMapper extends BaseMapperX<FileDO> {
     /**
      * 基于 Path 获取文件
      * 实际上，是基于 ID 查询
-     * 由于前端使用 <img /> 的方式获取图片，所以需要忽略租户的查询
      *
      * @param path 路径
      * @return 文件
      */
-    @InterceptorIgnore(tenantLine = "true")
     default FileDO selectByPath(String path) {
         return selectById(path);
     }
